chore: rename/restructure several components
